Implement grid view
