+++ src/utility/format-slot.js
@@ -1,0 +1,9 @@
+function formatSlotFromString(datetime) {
+  return new Date(datetime).toLocaleDateString('en-US', {
+    weekday: 'short',
+    hour: 'numeric',
+    minute: 'numeric'
+  });
+}
+
+export { formatSlotFromString };
